+++ analysis/filter/StructureModifier/src/main/kotlin/de/maibornwolff/codecharta/filter/structuremodifier/NodeRemover.kt
@@ -1,5 +1,6 @@
 package de.maibornwolff.codecharta.filter.structuremodifier
 
+import de.maibornwolff.codecharta.model.AttributeDescriptor
 import de.maibornwolff.codecharta.model.AttributeType
 import de.maibornwolff.codecharta.model.BlacklistItem
 import de.maibornwolff.codecharta.model.Edge
@@ -23,8 +24,9 @@ class NodeRemover(private val project: Project) {
             removeNodes(pathSegments),
             removeEdges(paths),
             copyAttributeTypes(),
+            copyAttributeDescriptors(),
             removeBlacklistItems(paths)
-        ).build()
+        ).build(true)
     }
 
     private fun filterNodes(path: List<String>, node: MutableNode): MutableNode {
@@ -53,6 +55,10 @@ class NodeRemover(private val project: Project) {
             mergedAttributeTypes[it.key] = it.value
         }
         return mergedAttributeTypes.toMutableMap()
+    }
+
+    private fun copyAttributeDescriptors(): MutableMap<String, AttributeDescriptor> {
+        return project.attributeDescriptors.toMutableMap()
     }
 
     private fun removeBlacklistItems(paths: Array<String>): MutableList<BlacklistItem> {